Initial import
This is the initial import of the Backbase Widget Factory. This is a
google chrome packaged application that can be downloaded and run from
the desktop.

The applications primary purpose is to help developers getting started
with Backbase a tool that assists in the creation, and registration of
Backbase Portal Assets (widgets, page templates).